requirements.txt
streamlit
numpy
requests
pybase64
pillow